web-app/backend/requirements.txt
fastapi>=0.110.0
uvicorn[standard]>=0.27.0
python-multipart>=0.0.9
python-dotenv>=1.0.0
opencv-python-headless>=4.9.0
numpy>=1.23.0
tensorflow>=2.10.0
scikit-learn>=1.1.0
pydantic>=2.0.0
